Supporting inherting of tests from abstract classes, allowing for common test suites for interface based tests
